Handle `Command.Poll` while subscribed only (#414)
* Handle `IllegalStateException` only if `subscribe` hasn't been called

`KafkaConsumer.poll` throws `IllegalStateException` for multiple
reasons, but we treat all of them as if they were only when no call to
`subscribe` has been made yet.

This is fixed by adding a guard to ensure we're on the mentioned
scenario.

* Track `Subscription` reference for ignoring `IllegalStateException`

Keep a reference for given `Subscription` applied and ignore
`IllegalStateException` only if type of `Subscription.Pattern` or not
subscribed yet.

* Resurrect comment about `IllegalStateException` on no-subscription

Co-authored-by: svroonland <svroonland@users.noreply.github.com>